+++ wire-c/wire-server/wire-listener.h
@@ -14,13 +14,13 @@ typedef struct
     char *packet;
 } ProtocolPacket;
 
-typedef int (*net_reader) (int socket, char *buffer);
+typedef ssize_t (*net_reader) (int socket, char *buffer);
 
 int handleRequest(char *buffer, struct wire_context *context);
 char *handle_callback(wire_feature_callback callback, struct wire_context *context);
-int getRequest(net_reader reader, int socket, char *buffer, size_t len);
+ssize_t getRequest(net_reader reader, int socket, char *buffer, ssize_t len);
 
-int getNetworkByte(int socket, char *buffer);
+ssize_t getNetworkByte(int socket, char *buffer);
 void cleanup(int socket1, int socket2);
 int acceptConnection(int sockfd);
 int makeSocket(int port);
